\name{brain}
\alias{brain}
\docType{data}
\title{Allometry}
\description{
  Allometry data from \cite{Van Aelst et al (2006)}
}
\usage{data(brain)}
\format{
  A matrix with 282 observations on the following 2 variables:
  \describe{
    \item{\code{BrainWeight.g.}}{a numeric vector}
    \item{\code{OlfactoryBulbs.ml.}}{a numeric vector}
  }
}
  
\references{
  Van Aelst, S. and Wang, X. and Zamar, R. and Zhu, R. (2006)
  \sQuote{Linear Grouping Using Orthogonal Regression},
  \emph{Computational Statistics \& Data Analysis} \bold{50}, 1287--1312.
}

\keyword{datasets}
