\name{corridorWalls}
\alias{corridorWalls}
\docType{data}
\title{Corridor Walls data}
\description{
  
  Corridor Walls data from Garcia-Escudero et al. (2008).  A laser
  device is introduced in a corridor of an office building and throws a
  laser ray which touches a point from the object found at the end of
  its trajectory. The device produces a three dimensional measurement of
  the placement of that point with respect to a fixed reference system.

}
\usage{data(corridorWalls)}
\format{
  A matrix with  11710 observations on the following 3 variables (3-dimensional coordinates).
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{z}}{a numeric vector}
  }
}

\references{

  Garcia-Escudero, L.A., Gordaliza, A., San Martin, R., Van Aelst,
  S. and Zamar, R.H. (2008) \sQuote{Robust linear clustering}. To appear
  in \emph{Journal of the Royal Statistical Society, Series B} (accepted
  June, 2008)

}

\keyword{datasets}
