\name{ob}
\alias{ob}
\docType{data}
\title{Allometry Data}
\description{
  Allometry data from \cite{Van Aelst et al (2006)}
}
\usage{data(ob)}
\format{
  A data frame with 83 observations on the following 3 variables.
  \describe{
    \item{\code{Group}}{a factor with levels \code{a} \code{c}
      \code{h} \code{H} \code{i} \code{m} \code{p} \code{t}}
    \item{\code{BrainWeight.g.}}{a numeric vector}
    \item{\code{OlfactoryBulbs.ml.}}{a numeric vector}
  }
}

\references{
    Van Aelst, S. and Wang, X. and Zamar, R. and Zhu, R. (2006)
  \sQuote{Linear Grouping Using Orthogonal Regression},
  \emph{Computational Statistics \& Data Analysis} \bold{50}, 1287--1312.
}

\keyword{datasets}
